+++ app/src/renderer/components/common/ActionModal.vue
@@ -56,22 +56,21 @@
           />
         </tm-form-group>
       </div>
+
       <div v-else-if="step === `sign`" class="action-modal-form">
-        <hardware-state
-          v-if="sending"
-          :loading="true"
-          value="Waiting for signature on app"
-        />
-        <hardware-state
-          v-else
-          icon="usb"
-          value="Please unlock the Cosmos app on your Ledger Nano&nbsp;S"
-        />
+        <hardware-state icon="usb" :loading="sending ? true : false">
+          {{
+            sending
+              ? `Please verify and sign the transaction on your Ledger`
+              : `Please plug in your Ledger&nbsp;Nano&nbsp;S and open the Cosmos app`
+          }}
+        </hardware-state>
       </div>
+
       <div class="action-modal-footer">
         <slot name="action-modal-footer">
           <tm-form-group class="action-modal-group">
-            <div class="action-modal-footer">
+            <div>
               <tm-btn
                 v-if="!session.signedIn"
                 value="Go to Sign In"
@@ -81,7 +80,7 @@
               />
               <tm-btn
                 v-else-if="sending"
-                value="Sending..."
+                :value="step === `sign` ? `Waiting for Ledger` : `Sending...`"
                 disabled="disabled"
                 color="primary"
               />
@@ -102,7 +101,7 @@
               <tm-btn
                 v-else-if="selectedSignMethod === `ledger` && step === `sign`"
                 color="primary"
-                value="Sign"
+                value="Next"
                 @click.native="validateChangeStep"
               />
               <tm-btn
@@ -222,6 +221,7 @@ export default {
       this.$emit(`close`)
     },
     goToSession() {
+      this.close()
       this.$store.commit(`setSessionModalView`, `welcome`)
       this.$store.commit(`toggleSessionModal`, true)
     },
@@ -285,11 +285,11 @@ export default {
   flex-direction: column;
   justify-content: space-between;
   right: 2rem;
-  padding: 3rem;
+  padding: 1.5rem 1.5rem 2rem 1.5rem;
   position: fixed;
   bottom: 0;
   width: 100%;
-  max-width: 664px;
+  max-width: 564px;
   z-index: var(--z-modal);
   border-top-left-radius: 0.25rem;
   border-top-right-radius: 0.25rem;
@@ -299,7 +299,7 @@ export default {
 .action-modal-header {
   align-items: center;
   display: flex;
-  padding-bottom: 2rem;
+  padding-bottom: 1.5rem;
 }
 
 .action-modal-atom {
@@ -335,18 +335,22 @@ export default {
 
 .action-modal-form .tm-form-group {
   display: block;
-  padding: 0.5rem 0 1rem;
+  padding: 0.75rem 0;
 }
 
 .action-modal-footer {
   display: flex;
   justify-content: flex-end;
-  padding: 2rem 0 0;
+  padding: 1.5rem 0 1rem;
+}
+
+.action-modal-footer .tm-form-group {
+  padding: 0;
 }
 
 .submission-error {
   position: absolute;
-  right: 3rem;
+  left: 1.5rem;
   bottom: 1rem;
 }