+++ test/unit/specs/components/governance/ModalDeposit.spec.js
@@ -50,6 +50,17 @@ describe(`ModalDeposit`, () => {
     expect(wrapper.vm.$refs.actionModal.open).toHaveBeenCalled()
   })
 
+  it(`clears on close`, () => {
+    wrapper.setData({ amount: 100000000000 })
+    // produce validation error as amount is too high
+    wrapper.vm.$v.$touch()
+    expect(wrapper.vm.$v.$error).toBe(true)
+
+    wrapper.vm.clear()
+    expect(wrapper.vm.$v.$error).toBe(false)
+    expect(wrapper.vm.amount).toBe(0)
+  })
+
   describe(`validation`, () => {
     describe(`fails`, () => {
       it(`with default values`, () => {
@@ -99,7 +110,7 @@ describe(`ModalDeposit`, () => {
           {
             amount: [
               {
-                amount: `100000000`,
+                amount: `100000000.0000000`,
                 denom: `stake`
               }
             ],