+++ test/unit/specs/components/staking/UndelegationModal.spec.js
@@ -58,6 +58,18 @@ describe(`UndelegationModal`, () => {
     expect(wrapper.vm.$refs.actionModal.open).toHaveBeenCalled()
   })
 
+  it(`clears on close`, () => {
+    wrapper.setData({ selectedIndex: 1, amount: 10000000000000 })
+    // produce validation error as amount is too high
+    wrapper.vm.$v.$touch()
+    expect(wrapper.vm.$v.$error).toBe(true)
+
+    wrapper.vm.clear()
+    expect(wrapper.vm.$v.$error).toBe(false)
+    expect(wrapper.vm.selectedIndex).toBe(0)
+    expect(wrapper.vm.amount).toBe(null)
+  })
+
   describe(`only submits on correct form`, () => {
     describe(`validates`, () => {
       it(`to false with default values`, () => {
@@ -71,35 +83,37 @@ describe(`UndelegationModal`, () => {
     })
   })
 
-  describe(`Undelegate`, () => {
+  describe(`submitForm`, () => {
     it(`submits undelegation`, async () => {
-      wrapper.vm.$store.dispatch = jest.fn()
-      wrapper.vm.$store.commit = jest.fn()
+      const $store = {
+        dispatch: jest.fn(),
+        commit: jest.fn()
+      }
+      const validator = {
+        operator_address: `cosmosvaloper1address`,
+      }
 
       wrapper.setData({ amount: 4.2 })
-      await wrapper.vm.submitForm(`local`, `1234567890`)
+      await UndelegationModal.methods.submitForm.call(
+        { $store, amount: 4.2, denom: `atom`, validator },
+        `local`, `1234567890`
+      )
 
-      expect(wrapper.vm.$store.dispatch.mock.calls).toEqual([
-        [
-          `submitUnbondingDelegation`,
-          {
-            amount: -42000000,
-            validator,
-            submitType: `local`,
-            password: `1234567890`
-          }
-        ]
-      ])
+      expect($store.dispatch).toHaveBeenCalledWith(`submitUnbondingDelegation`,
+        {
+          amount: -42000000,
+          validator,
+          submitType: `local`,
+          password: `1234567890`
+        }
+      )
 
-      expect(wrapper.vm.$store.commit.mock.calls).toEqual([
-        [
-          `notify`,
-          {
-            body: `You have successfully undelegated 4.2 STAKEs.`,
-            title: `Successful undelegation!`
-          }
-        ]
-      ])
+      expect($store.commit).toHaveBeenCalledWith(`notify`,
+        {
+          body: `You have successfully undelegated 4.2 atoms.`,
+          title: `Successful undelegation!`
+        }
+      )
     })
   })
 })